refactor(admin, auth, dto): add data class, move to a new package, implement search by queries
