+++ src/main/kotlin/ecommerce/dto/cart/CartResponse.kt
@@ -2,7 +2,7 @@ package ecommerce.dto.cart
 
 import ecommerce.dto.cartItem.CartItemResponse
 
-class CartResponse(
+data class CartResponse(
     val id: Long,
     val memberId: Long,
     val items: List<CartItemResponse>,
