+++ src/main/kotlin/ecommerce/dto/member/AuthResponse.kt
@@ -1,6 +1,6 @@
 package ecommerce.dto.member
 
-class AuthResponse(
+data class AuthResponse(
     val success: Boolean,
     val message: String,
     val token: String,
