+++ src/main/kotlin/ecommerce/repository/reportDTO/ProductCartCountDto.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto.report
+package ecommerce.repository.reportDTO
 
 import java.time.LocalDateTime
